+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 - Show thumbnail in chatlist summary of image, sticker and webxdc messages
 - add webxdc api `sendToChat` #3240
+- add webxdc api `importFiles`
 
 ### Changed
 - exclude more unused files from installation package